add back in vtable conversion removed by mistake
