+++ app/save-and-restore/app/src/main/java/org/phoebus/applications/saveandrestore/ui/Utilities.java
@@ -17,8 +17,6 @@
  */
 package org.phoebus.applications.saveandrestore.ui;
 
-import org.epics.pva.data.*;
-import org.epics.pva.data.nt.PVATable;
 import org.epics.util.array.*;
 import org.epics.util.number.*;
 import org.epics.util.text.NumberFormats;
@@ -1300,65 +1298,4 @@ public final class Utilities {
         return false;
     }
 
-    /**
-     * @param name Data item name
-     * @param object The object subject to conversion
-     * @return Converted object
-     */
-    public static Object toPVArrayType(String name, Object object) {
-        if (object instanceof ListBoolean) {
-            ListBoolean listBoolean = (ListBoolean) object;
-            boolean[] booleans = new boolean[listBoolean.size()];
-            for (int i = 0; i < listBoolean.size(); i++) {
-                booleans[i] = listBoolean.getBoolean(i);
-            }
-            return new PVABoolArray(name, booleans);
-        } else if (object instanceof ListNumber) {
-            ListNumber listNumber = (ListNumber) object;
-            if (object instanceof ArrayByte || object instanceof ArrayUByte) {
-                byte[] bytes = new byte[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    bytes[i] = listNumber.getByte(i);
-                }
-                return new PVAByteArray(name, object instanceof ArrayUByte, bytes);
-            } else if (object instanceof ArrayShort || object instanceof ArrayUShort) {
-                short[] shorts = new short[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    shorts[i] = listNumber.getShort(i);
-                }
-                return new PVAShortArray(name, object instanceof ArrayUShort, shorts);
-            } else if (object instanceof ArrayInteger || object instanceof ArrayUInteger) {
-                int[] ints = new int[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    ints[i] = listNumber.getInt(i);
-                }
-                return new PVAIntArray(name, object instanceof ArrayUInteger, ints);
-            } else if (object instanceof ArrayLong || object instanceof ArrayULong) {
-                long[] longs = new long[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    longs[i] = listNumber.getLong(i);
-                }
-                return new PVALongArray(name, object instanceof ArrayULong, longs);
-            } else if (object instanceof ArrayFloat) {
-                float[] floats = new float[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    floats[i] = listNumber.getFloat(i);
-                }
-                return new PVAFloatArray(name, floats);
-            } else if (object instanceof ArrayDouble) {
-                double[] doubles = new double[listNumber.size()];
-                for (int i = 0; i < listNumber.size(); i++) {
-                    doubles[i] = listNumber.getDouble(i);
-                }
-                return new PVADoubleArray(name, doubles);
-            } else {
-                throw new IllegalArgumentException("Conversion of type " + object.getClass().getCanonicalName() + " not supported");
-            }
-        } else { // Assume this always is for string arrays
-            Collection<String> list = (Collection<String>) object;
-            String[] strings = new String[list.size()];
-            strings = list.toArray(strings);
-            return new PVAStringArray(name, strings);
-        }
-    }
 }
